Created the folders and the notebook
We cannot add the entire dataset to the repo due to the limited storage of the free account. Another option is github LFS but we need to figure out how to use it or we could simply download the data locally on our devices or use another cloud storage platform.